+++ src/render/tui/draw/progress.rs
@@ -1,4 +1,5 @@
 use crate::{
+    progress::{self, Progress, Step, Value},
     render::tui::{
         draw::State,
         utils::{
@@ -8,11 +9,7 @@ use crate::{
         InterruptDrawInfo,
     },
     time::format_now_datetime_seconds,
-    tree::{
-        self,
-        progress::{self, Step, Value},
-        Key, Progress,
-    },
+    tree::{self, Key},
     unit,
 };
 use humantime::format_duration;