+++ src/tree/throughput.rs
@@ -1,4 +1,7 @@
-use crate::{tree, tree::progress, tree::Progress, unit};
+use crate::{
+    progress::{self, Progress},
+    tree, unit,
+};
 use std::time::{Duration, SystemTime};
 
 const THROTTLE_INTERVAL: Duration = Duration::from_secs(1);