+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Breaking
 
 * In `Progress::init(max, unit)` , `unit` now is `Option<Unit>`, existing code can be transformed using `progress.init(None, Some("label".into()))`.
+* moved `tree::progress` into `progress` and renamed `Value` into `Task` and `Progress` into `Value`
 * **removed** `deep_eq()` method in `Root` tree
 * tui engine option `redraw_only_on_state_change` was removed without substitute
 * **Move and rename**